scripts/requirements.txt
jsonpath_ng
requests
azure-identity
anytree